More rearranging and documentation updates.
