+++ hoomd-interaction/src/single.rs
@@ -4,20 +4,21 @@
 /*! Implement Single
 */
 
-use crate::{SiteEnergy, TotalEnergy};
-use hoomd_microstate::Microstate;
+use crate::{DeltaEnergyOne, SiteEnergy, TotalEnergy};
+use hoomd_microstate::{Body, Microstate, Transform, boundary::Boundary, property::Position};
 
-/** Compute system properties given a [`SiteEnergy`]
+/** Compute system properties from external fields.
 
-[`Single`] is a newtype that provides a single implementation for system
-properties, like [`TotalEnergy`], for all types that implement [`SiteEnergy`].
+Given an inner type that implements [`SiteEnergy`], [`Single`] represents:
 
-Use types that implement [`SiteEnergy`], such as one from
-[`external`](crate::external) or your own custom type, directly when you
-only need to call `site_energy`. Wrap the type in `Single` to use it with MC
-simulations or to compute the total energy.
+```math
+U_\mathrm{total} = \sum_{i=0}^{N-1} U\left( s_i \right)
+```
+where $`s_i`$ is the full set of site properties for site i.
 
-TODO: Reword this when `Single` also implements `SiteForce`.
+For the inner type, use one from [`external`] or your own custom type.
+
+[`external`]: crate::external
 
 # Example
 
@@ -52,6 +53,28 @@ where
     The sum over sites differs from HOOMD-blue where external energies are
     evaluated only at the body centers. In general, hoomd-rs interactions apply
     to sites. Use a custom implementation to compute energies over body centers.
+
+    # Example
+
+    ```
+    use hoomd_interaction::{Single, TotalEnergy, external::Linear};
+    use hoomd_microstate::{Microstate, Body, property::Point};
+    use hoomd_vector::Cartesian;
+
+    # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let mut microstate = Microstate::new();
+    microstate.extend_bodies([Body::point(Cartesian::from([1.0, 0.0])),
+                              Body::point(Cartesian::from([-1.0, 2.0]))])?;
+
+    let linear = Single(Linear{ alpha: 1.0,
+        plane_origin: Cartesian::default(),
+        plane_normal: [0.0, 1.0].try_into()? });
+
+    let total_energy = linear.total_energy(&microstate);
+    assert_eq!(total_energy, 2.0);
+    # Ok(())
+    # }
+    ```
     */
     #[inline]
     fn total_energy(&self, microstate: &Microstate<B, S, C>) -> f64 {
@@ -59,6 +82,63 @@ where
             .sites()
             .iter()
             .fold(0.0, |total, s| total + self.0.site_energy(&s.properties))
+    }
+}
+
+/** Evaluate the change in energy due to functions that act on single sites.
+
+# Example
+
+```
+use hoomd_interaction::{DeltaEnergyOne, Single, external::Linear};
+use hoomd_microstate::{Microstate, Body, property::Point};
+use hoomd_vector::Cartesian;
+
+# fn main() -> Result<(), Box<dyn std::error::Error>> {
+let mut microstate = Microstate::new();
+microstate.add_body(Body::point(Cartesian::from([0.0, 0.0])))?;
+
+let linear = Single(Linear{ alpha: 1.0,
+    plane_origin: Cartesian::default(),
+    plane_normal: [0.0, 1.0].try_into()? });
+
+let delta_energy = linear.delta_energy_one(&microstate, 0,
+    &Body::point([0.0, -1.0].into()));
+assert_eq!(delta_energy, -1.0);
+# Ok(())
+# }
+```
+*/
+impl<V, B, S, C, E> DeltaEnergyOne<B, S, C> for Single<E>
+where
+    E: SiteEnergy<S>,
+    B: Transform<S>,
+    S: Position<Vector = V>,
+    C: Boundary<V, B, S>,
+{
+    #[inline]
+    fn delta_energy_one(
+        &self,
+        initial_microstate: &Microstate<B, S, C>,
+        body_index: usize,
+        final_body: &Body<B, S>,
+    ) -> f64 {
+        let mut energy_final = 0.0;
+        for s in &final_body.sites {
+            match initial_microstate
+                .boundary()
+                .wrap_site(final_body.properties.transform(s))
+            {
+                Ok(wrapped_site) => energy_final += self.site_energy(&wrapped_site),
+                Err(_) => return f64::INFINITY,
+            }
+        }
+
+        let energy_initial = initial_microstate
+            .iter_body_sites(body_index)
+            .fold(0.0, |total, s| total + self.site_energy(&s.properties));
+
+        energy_final - energy_initial
     }
 }
 
@@ -75,9 +155,9 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use hoomd_microstate::boundary::Open;
+    use hoomd_microstate::boundary::{Open, Square};
     use hoomd_microstate::property::{Point, Position};
-    use hoomd_microstate::{Body, Microstate};
+    use hoomd_microstate::{Body, Microstate, MicrostateBuilder};
     use hoomd_vector::Cartesian;
     use rstest::*;
 
@@ -119,5 +199,45 @@ mod tests {
 
         assert_eq!(single.site_energy(&microstate.sites()[0].properties), 1.0);
         assert_eq!(single.site_energy(&microstate.sites()[1].properties), 2.0);
+    }
+
+    mod delta_energy {
+        use super::*;
+
+        struct Zero;
+
+        impl SiteEnergy<Point<Cartesian<2>>> for Zero {
+            fn site_energy(&self, _site_properties: &Point<Cartesian<2>>) -> f64 {
+                0.0
+            }
+        }
+
+        #[test]
+        fn site_outside() {
+            let square = Square {
+                l: 4.0
+                    .try_into()
+                    .expect("hard-coded constant should be positive"),
+            };
+
+            let body = Body {
+                properties: Point::new(Cartesian::from([0.0, 0.0])),
+                sites: [Point::new(Cartesian::from([1.0, 0.0]))].into(),
+            };
+            let mut final_body = body.clone();
+            final_body.properties.position[0] = 1.0;
+
+            let microstate = MicrostateBuilder::with_boundary(square)
+                .bodies([body])
+                .try_build()
+                .expect("the hard-coded bodies should be in the boundary");
+
+            let energy = Single(Zero);
+
+            assert_eq!(
+                energy.delta_energy_one(&microstate, 0, &final_body),
+                f64::INFINITY
+            );
+        }
     }
 }
